requirements.txt
antlr4-python3-runtime==4.13.1
coverage==7.3.1
simplejson==3.19.1
typing==3.7.4.3
typing_extensions==4.8.0
pezzo==0.1.2
tiktoken==0.3.3
faiss-cpu
psycopg2-binary==2.9.9
openai
tiktoken
pandas
bs4
requests
scikit-learn
selenium
prompttools
gensim
sqlalchemy
python-dotenv
colorama
matplotlib